Updates index.html with a responsive header.
Adds a responsive header to index.html. I got problems trying to implement the fixed header that we discussed. So it's not part of the html and css for now. I also changed the background color to one of the lighter ones in the palette which works better with the black icons.